frontend/requirements.txt
PyYAML
celery<3.2
SQLAlchemy
marshmallow<2.0
hug==2.3.0
mock
coverage
nose
requests
python-magic
paramiko>2.0
fasteners
falcon-multipart
https://projects.unbit.it/downloads/uwsgi-lts.tar.gz
alembic
humanfriendly
psycopg2 --no-binary :all:
